[BUG] Null pointer dereference (#191)
* add checks and throw exceptions

* fix ci

* comments

* use log4j2

+++ src/main/java/com/amazonaws/kinesisvideo/java/service/JavaKinesisVideoServiceClient.java
@@ -35,6 +35,7 @@ import com.amazonaws.services.kinesisvideo.model.GetDataEndpointRequest;
 import com.amazonaws.services.kinesisvideo.model.GetDataEndpointResult;
 import com.amazonaws.services.kinesisvideo.model.TagStreamRequest;
 import com.amazonaws.services.kinesisvideo.model.TagStreamResult;
+import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
 import javax.annotation.Nonnull;
@@ -52,7 +53,7 @@ public final class JavaKinesisVideoServiceClient implements KinesisVideoServiceC
     private static final int RECEIVE_TIMEOUT_1HR = 60 * 60 * 1000;
     private static final String ABSOLUTE_TIMECODE = "ABSOLUTE";
     private static final String RELATIVE_TIMECODE = "RELATIVE";
-
+    private static final Logger logger = LogManager.getLogger(JavaKinesisVideoServiceClient.class);
     private final Logger log;
     private KinesisVideoClientConfiguration configuration;
 
@@ -117,6 +118,11 @@ public final class JavaKinesisVideoServiceClient implements KinesisVideoServiceC
 
         final KinesisVideoCredentials kinesisVideoCredentials = credentialsProvider.getCredentials();
 
+        if (kinesisVideoCredentials == null) {
+            logger.error("kinesisVideoCredentials is null");
+            return null;
+        }
+
         AWSCredentials credentials = null;
 
         if (kinesisVideoCredentials.getSessionToken() == null) {
@@ -169,6 +175,11 @@ public final class JavaKinesisVideoServiceClient implements KinesisVideoServiceC
                 try {
                     final KinesisVideoCredentials kinesisVideoCredentials = credentialsProvider.getCredentials();
 
+                    if (kinesisVideoCredentials == null) {
+                        logger.error("kinesisVideoCredentials must not be null while obtaining it from getCredentials");
+                        throw new IllegalArgumentException();
+                    }
+
                     if (kinesisVideoCredentials.getSessionToken() == null) {
                         awsCredentials = new AWSCredentials() {
                             @Override
@@ -199,7 +210,7 @@ public final class JavaKinesisVideoServiceClient implements KinesisVideoServiceC
                             }
                         };
                     }
-                } catch (final KinesisVideoException e) {
+                } catch (final KinesisVideoException | IllegalArgumentException e) {
                     log.error("Getting credentials threw an exception.", e);
                     awsCredentials = null;
                 }
@@ -339,21 +350,27 @@ public final class JavaKinesisVideoServiceClient implements KinesisVideoServiceC
 
         final StreamDescription streamDescription = describeStream(streamName, timeoutInMillis, credentialsProvider);
 
-        final DeleteStreamRequest deleteStreamRequest = new DeleteStreamRequest()
-                .withStreamARN(streamDescription.getStreamArn())
-                .withCurrentVersion(streamDescription.getUpdateVersion());
-
-        log.debug("calling delete stream: {}", deleteStreamRequest.toString());
-
         final DeleteStreamResult deleteStreamResult;
+
         try {
+            if (streamDescription == null) {
+                log.error("Stream description must not be null");
+                throw new IllegalArgumentException();
+            }
+
+            final DeleteStreamRequest deleteStreamRequest = new DeleteStreamRequest()
+                    .withStreamARN(streamDescription.getStreamArn())
+                    .withCurrentVersion(streamDescription.getUpdateVersion());
+
+            log.debug("calling delete stream: {}", deleteStreamRequest.toString());
             deleteStreamResult = serviceClient.deleteStream(deleteStreamRequest);
+            log.debug("delete stream result: {}", deleteStreamResult.toString());
         } catch (final AmazonClientException e) {
             log.error("Service call failed.", e);
             throw new KinesisVideoException(e);
+        } catch (final IllegalArgumentException e) {
+            log.error("Stream description null.", e);
         }
-
-        log.debug("delete stream result: {}", deleteStreamResult.toString());
     }
 
     @Override
